\name{prewtcl-package}
\alias{prewtcl-package}
\alias{prewtcl}
\docType{package}
\title{
Clustering based on the preweighted sparse k-means method
}
\description{
Implements the preweighted sparse clustering method of Gaynor and Bair.
}
\details{
\tabular{ll}{
Package: \tab prewtcl\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-22\cr
License: \tab GPL-3 \cr
}
The primary function is KMeansSparseCluster2, which performs preweighted sparse clustering. Weights and the tuning parameter for this function are chosen by sig.features and KMeansSparseCluster.permute2.
}
\author{
Sheila Gaynor and Eric Bair

Maintainer: <ebair@email.unc.edu>

}
\references{
Gaynor, Sheila and Bair, Eric. (2013) "Identification of biologically relevant subtypes via preweighted sparse clustering".  arXiv:1304.3760v2.
}
\keyword{ package }
